// src/commons/connector/yaci/types/drep-page.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yaci Store API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { DRepRegistration } from "./drep-registration";
/**
 *
 *
 * @export
 * @interface DRepPage
 */
export interface DRepPage {
  /**
   * @type {number}
   * @memberof DRepPage
   */
  total?: number;

  /**
   * @type {number}
   * @memberof DRepPage
   */
  totalPages?: number;

  /**
   * @type {Array<DRepRegistration>}
   * @memberof DRepPage
   */
  dreps?: Array<DRepRegistration>;
}